refactor(core): consolidate environment APIs under environment domain
- Move all environment-related APIs into the environment folder
- Remove obsolete environment APIs
- Adopt domain-based development structure
