+++ npm/react/libs/core/src/environment/use-environment.ts
@@ -1,4 +1,4 @@
 import { createStoreHook } from '../store/axiom-store';
-import { environmentStore } from '../store/environment-store';
+import { environmentStore } from './environment-store';
 
 export const useEnvironment = createStoreHook(environmentStore);
