--- npm/react/libs/core/src/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useEnvironment } from './use-environment';
