+++ npm/react/libs/core/src/index.ts
@@ -1,6 +1,6 @@
 export * from './auth/index';
 export * from './config/index';
-export * from './hooks/index';
+export * from './environment/index';
 export * from './http/index';
 export * from './initializer/index';
 export * from './localization/index';
