+++ npm/react/libs/core/src/models/index.ts
@@ -1,4 +1,3 @@
 export { type ApplicationInitializer, InitializerError } from './application';
-export type { Endpoint, Environment } from './environment';
 export type { Awaitable, Provider } from './common';
 export type { OAuth } from './oauth';
