+++ npm/react/libs/core/src/store/index.ts
@@ -1,2 +1,1 @@
 export { createStore, createStoreHook } from './axiom-store';
-export { environmentStore } from './environment-store';
